requirements.txt
torch>=2.1
pyyaml>=6.0
tqdm>=4.66
numpy>=1.23
# 可选：有网时可用
transformers>=4.44
datasets>=2.20
tokenizers>=0.15
